requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
requests==2.31.0
opencv-python==4.8.1.78
timm==0.9.2
scipy==1.11.3 